requirements.txt
azure-functions
sklearn
spacy
nltk
pandas
gensim